Add permissions for deleting note

+++ app/routes/users+/$username_+/notes.$noteId.tsx
@@ -17,7 +17,7 @@ import { ErrorList } from '~/components/forms'
 import { Button } from '~/components/ui/button'
 import { Icon } from '~/components/ui/icon'
 import { StatusButton } from '~/components/ui/status-button'
-import { requireUser } from '~/utils/auth.server'
+import { getUserId, requireUser } from '~/utils/auth.server'
 import { validateCSRF } from '~/utils/csrf.server'
 import { prisma } from '~/utils/db.server'
 import { getNoteImgSrc, invariantResponse, useIsPending } from '~/utils/misc'
@@ -25,7 +25,8 @@ import { toastSessionStorage } from '~/utils/toast.server'
 import { useOptionalUser } from '~/utils/user'
 import { type loader as notesLoader } from './notes'
 
-export async function loader({ params }: LoaderFunctionArgs) {
+export async function loader({ params, request }: LoaderFunctionArgs) {
+	const userId = await getUserId(request)
 	const note = await prisma.note.findUnique({
 		where: { id: params.noteId },
 		select: {
@@ -48,9 +49,31 @@ export async function loader({ params }: LoaderFunctionArgs) {
 	const date = new Date(note.updatedAt)
 	const timeAgo = formatDistanceToNow(date)
 
+	// this query is a little tricky if you're unfamiliar with Prisma so make
+	// sure to check the example in the instructions.
+	// get the permission here. If the userId does not exist, don't bother,
+	// as the permission should just be null. If it does though, get the
+	// permission where "some" of the permission's roles have "some" users with the userId
+	// https://www.prisma.io/docs/reference/api-reference/prisma-client-reference#some
+	// you have the note.ownerId, so you can use that to decide whether you
+	// should be looking for "access: 'own'" or "access: 'any'".
+
+	const permission = userId
+		? await prisma.permission.findFirst({
+				select: { id: true },
+				where: {
+					roles: { some: { users: { some: { id: userId } } } },
+					entity: 'note',
+					action: 'delete',
+					access: note.ownerId === userId ? 'own' : 'any',
+				},
+			})
+		: null
+
 	return json({
 		note,
 		timeAgo,
+		canDelete: Boolean(permission),
 	})
 }
 
@@ -61,9 +84,6 @@ const DeleteFormSchema = z.object({
 
 export async function action({ params, request }: ActionFunctionArgs) {
 	const user = await requireUser(request)
-	invariantResponse(user.username === params.username, 'Not authorized', {
-		status: 403,
-	})
 
 	const formData = await request.formData()
 	await validateCSRF(formData, request.headers)
@@ -80,10 +100,32 @@ export async function action({ params, request }: ActionFunctionArgs) {
 	const { noteId } = submission.value
 
 	const note = await prisma.note.findFirst({
-		select: { id: true, owner: { select: { username: true } } },
-		where: { id: noteId, ownerId: user.id },
+		select: { id: true, ownerId: true, owner: { select: { username: true } } },
+		where: { id: noteId },
 	})
 	invariantResponse(note, 'Not found', { status: 404 })
+
+	const permission = await prisma.permission.findFirst({
+		select: { id: true },
+		where: {
+			roles: { some: { users: { some: { id: user.id } } } },
+			entity: 'note',
+			action: 'delete',
+			access: note.ownerId === user.id ? 'own' : 'any',
+		},
+	})
+
+	// if there is no permission, then throw a response with a 403 status code
+	// which you can handle in the error boundary below
+	if (!permission) {
+		throw json(
+			{
+				error: 'Unauthorized',
+				message: `Unauthorized: requires permission delete:note`,
+			},
+			{ status: 403 },
+		)
+	}
 
 	await prisma.note.delete({ where: { id: note.id } })
 
@@ -109,13 +151,13 @@ export default function NoteRoute() {
 	const user = useOptionalUser()
 	const isOwner = user?.id === data.note.ownerId
 
-	const canDelete = true
+	const canDelete = data.canDelete
 	const displayBar = canDelete || isOwner
 
 	return (
 		<div className="absolute inset-0 flex flex-col px-10">
 			<h2 className="mb-2 pt-12 text-h2 lg:mb-6">{data.note.title}</h2>
-			<div className="overflow-y-auto pb-24">
+			<div className={`${displayBar ? 'pb-24' : 'pb-12'} overflow-y-auto`}>
 				<ul className="flex flex-wrap gap-5 py-5">
 					{data.note.images.map(image => (
 						<li key={image.id}>
@@ -141,7 +183,7 @@ export default function NoteRoute() {
 						</Icon>
 					</span>
 					<div className="grid flex-1 grid-cols-2 justify-end gap-2 min-[525px]:flex md:gap-4">
-						<DeleteNote id={data.note.id} />
+						{canDelete ? <DeleteNote id={data.note.id} /> : null}
 						<Button
 							asChild
 							className="min-[525px]:max-md:aspect-square min-[525px]:max-md:px-0"
@@ -226,6 +268,7 @@ export function ErrorBoundary() {
 	return (
 		<GeneralErrorBoundary
 			statusHandlers={{
+				403: () => <p>You are not allowed to do that</p>,
 				404: ({ params }) => (
 					<p>No note with the id "{params.noteId}" exists</p>
 				),
